fixing unsigned/signed int comparison warnings
git-svn-id: https://svn.lcsr.jhu.edu/mechatronics/trunk/Software@106 95aa9265-a27f-408e-9981-efeafd1b3a66

+++ tests/qladisp.cpp
@@ -49,7 +49,7 @@ int main(int argc, char** argv)
     int board2 = BoardIO::MAX_BOARDS;
 
     int args_found = 0;
-    for (i = 1; i < argc; i++) {
+    for (i = 1; i < (unsigned int)argc; i++) {
         if ((argv[i][0] == '-') && (argv[i][1] == 'p')) {
             port = atoi(argv[i]+2);
             std::cerr << "Selecting port " << port << std::endl;
